Move some utility methods to separate package
