+++ package.json
@@ -49,7 +49,8 @@
     "vitest": "^1.5.0"
   },
   "dependencies": {
-    "@sveltia/ui": "^0.12.0",
+    "@sveltia/ui": "^0.12.1",
+    "@sveltia/utils": "^0.1.0",
     "base32-encode": "^2.0.0",
     "deepmerge": "^4.3.1",
     "fast-deep-equal": "^3.1.3",
