+++ src/lib/components/assets/toolbar/download-assets-button.svelte
@@ -1,8 +1,9 @@
 <script>
-  import { Alert, Button, Toast, sleep } from '@sveltia/ui';
+  import { Alert, Button, Toast } from '@sveltia/ui';
+  import { saveFile } from '@sveltia/utils/file';
+  import { sleep } from '@sveltia/utils/misc';
   import { _ } from 'svelte-i18n';
   import { getAssetBlob } from '$lib/services/assets';
-  import { saveFile } from '$lib/services/utils/files';
 
   /**
    * @type {Asset[]}
