+++ src/lib/components/contents/details/toolbar.svelte
@@ -15,6 +15,7 @@
     Toast,
     Toolbar,
   } from '@sveltia/ui';
+  import { truncate } from '@sveltia/utils/string';
   import equal from 'fast-deep-equal';
   import { _ } from 'svelte-i18n';
   import { goBack, goto } from '$lib/services/app/navigation';
@@ -31,7 +32,6 @@
   import { getAssociatedAssets } from '$lib/services/contents/entry';
   import { defaultI18nConfig, getLocaleLabel } from '$lib/services/contents/i18n';
   import { formatSummary } from '$lib/services/contents/view';
-  import { truncate } from '$lib/services/utils/strings';
 
   let showDuplicateToast = false;
   let showValidationToast = false;
