+++ src/lib/components/contents/details/widgets/object/object-preview.svelte
@@ -4,9 +4,9 @@
   @see https://decapcms.org/docs/widgets/#object
 -->
 <script>
+  import { waitForVisibility } from '@sveltia/utils/element';
   import FieldPreview from '$lib/components/contents/details/preview/field-preview.svelte';
   import { entryDraft } from '$lib/services/contents/editor';
-  import { waitForVisibility } from '$lib/services/utils/misc';
 
   /**
    * @type {LocaleCode}
