+++ src/lib/services/backends/github.js
@@ -1,3 +1,5 @@
+import { getBase64 } from '@sveltia/utils/file';
+import { stripSlashes } from '@sveltia/utils/string';
 import mime from 'mime';
 import { _ } from 'svelte-i18n';
 import { get } from 'svelte/store';
@@ -6,9 +8,7 @@ import { createCommitMessage } from '$lib/services/backends/shared/commits';
 import { fetchAndParseFiles, repositoryProps } from '$lib/services/backends/shared/data';
 import { siteConfig } from '$lib/services/config';
 import { user } from '$lib/services/user';
-import { getBase64 } from '$lib/services/utils/files';
 import { sendRequest } from '$lib/services/utils/networking';
-import { stripSlashes } from '$lib/services/utils/strings';
 
 const backendName = 'github';
 const label = 'GitHub';
