+++ src/lib/services/prefs.js
@@ -1,7 +1,7 @@
+import LocalStorage from '@sveltia/utils/local-storage';
 import equal from 'fast-deep-equal';
 import { locale as appLocale, locales as appLocales } from 'svelte-i18n';
 import { get, writable } from 'svelte/store';
-import LocalStorage from '$lib/services/utils/local-storage';
 
 const storageKey = 'sveltia-cms.prefs';
